move goal calculations to separate files by type
